Imagemetadata/requirements.txt
Pillow>=10.4.0
piexif>=1.1.3
